requirements.txt
appdirs==1.4.3
boto3==1.15.12
botocore==1.18.12
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
decorator==4.4.2
dgl-cu102==0.4.3.post2
dill==0.3.2
distlib==0.3.0
distro==1.4.0
future==0.18.2
gensim==3.8.3
graphviz==0.14.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jmespath==0.10.0
joblib==0.17.0
kiwisolver==1.2.0
lockfile==0.12.2
matplotlib==3.3.2
msgpack==0.6.2
networkx==2.5
numpy==1.19.2
packaging==20.3
pandas==1.1.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
pydot==1.4.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
six==1.14.0
smart-open==2.2.1
threadpoolctl==2.1.0
torch==1.5.1
torchvision==0.6.1
tqdm==4.50.0
urllib3==1.25.8
webencodings==0.5.1
